requirements.txt
alabaster
anyio==4.8.0
arrow==1.3.0
attrs==25.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bibtexparser==1.4.3
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
cssselect==1.2.0
cssutils==2.11.1
Deprecated==1.2.18
docutils==0.21.2
emails==0.6
exceptiongroup==1.2.2
fake-useragent==2.0.3
Flask==3.1.0
Gunicorn
flask_httpauth
python-dotenv
free-proxy==1.1.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.6.0
outcome==1.3.0.post0
packaging==24.2
premailer==3.10.0
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scholarly==1.7.11
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
swiftshadow==1.2.1
tomli==2.2.1
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
